Add limit of 40 characters to scoreboard lines
As per the gitlab issue:

Scoreboards seem to have a limit of 40 characters per entry that was not taken into account. If a server has the auctions.toggles.display-custom-name config setting enabled, and the item display name is longer than 40 characters (including color codes), bukkit throws an IllegalArgumentException.

To solve this, any scoreboard line greater than length 40 characters will be truncated to 40